Refactor code for increased readability
