+++ src/components/Explorer.jsx
@@ -1,56 +1,401 @@
-import React from 'react';
-import Container from 'react-bootstrap/Container';
-import Row from 'react-bootstrap/Row';
-import Col from 'react-bootstrap/Col';
-import Image from 'react-bootstrap/Image';
+import axios from 'axios';
+import Button from 'react-bootstrap/Button';
+import Form from 'react-bootstrap/Form';
+import { useState } from 'react';
 
+import CityInfo from './CityInfo';
+import ErrorModal from './ErrorModal';
+import '../css/userForm.css';
 
-class Explorer extends React.Component {
+const API_KEY = import.meta.env.VITE_LOCATIONIQ_API_KEY;
+
+/***************************************************** */
+const Explorer = () => {
+  // State hooks
+  const [latitude, setLatitude] = useState('');
+  const [longitude, setLongitude] = useState('');
+  const [cityName, setCityName] = useState('');
+  const [userInput, setUserInput] = useState('');
+  const [mapImageUrl, setMapImageUrl] = useState('');
+  const [errorResponse, setErrorResponse] = useState('');
+  const [errorResponseBody, setErrorResponseBody] = useState('');
+  const [show, setShow] = useState(false);
+
+  const handleUserInput = (event) => {
+    setUserInput(event.target.value);
+  }
   
-  constructor (props) {
+  const handleFormSubmit = (event) => {
+    event.preventDefault();
+    if (userInput) {
+      getLocationInfo(API_KEY, userInput);
+    }
+  };
 
-    super(props); // Activates React.Component
-    
+  // Function to fetch location info based on user input and API key
+  const getLocationInfo = async (API_KEY, city) => {
+    try {
+      // Fetch location data
+      const locationInfo = await fetchLocationData(API_KEY, city);
 
-  }
-
-  render () {
-
-     // Write props passed in from parent in one line, instead of 'this.props' everytime used
-     let { displayName, latitude, longitude, staticMapURL } = this.props
-
-    return (
+      // Update state based on location data
+      updateLocationState(locationInfo);
       
-      <main>
+    } catch (error) {
+      // Handle errors
+      handleLocationError(error);
+    }
+  };
 
-        <ul className='city-data'>
-          <li>{displayName}</li>
-            {latitude && <li>latitude: {latitude}</li>}
-            {longitude && <li>longitude: {longitude}</li>}
-        </ul>
+  // Function to make API request
+  const fetchLocationData = async (API_KEY, city) => {
+    // Make API request to get location info
+    const API = `https://us1.locationiq.com/v1/search?key=${API_KEY}&q=${city}&format=json`;
+    const response = await axios.get(API);
+    return response.data[0];
+  };
 
-        <Container>
-          <Row className='map-row'>
-            <Col xs={12} md={12} className='map-column'>
+  // Function to extract and set state from location data
+  const updateLocationState = (locationInfo) => {
+    // Extract data from response
+    const cityName = locationInfo.display_name;
+    const latitude = locationInfo.lat;
+    const longitude = locationInfo.lon;
 
-              {staticMapURL && (
-                <Image 
-                  src={staticMapURL} 
-                  alt={`map image of ${displayName}`} 
-                  roundedCircle
-                  className='map-image'
-                />
-              )}
-        
-            </Col>
-          </Row> 
-        </Container>
+    // Update state with data received
+    setCityName(cityName);
+    setLatitude(latitude);
+    setLongitude(longitude);
+    
+    // Set image url after latitude and longitude are recieved
+    const mapUrl = getMapURL(API_KEY, latitude, longitude);
+    setMapImageUrl(mapUrl);
+  };
 
-      </main>
-                
-    );
+  // Function to generate map URL based on latitude and longitude
+  const getMapURL = (API_KEY, latitude, longitude) => {
+    return`https://maps.locationiq.com/v3/staticmap?key=${API_KEY}&center=${latitude},${longitude}&zoom=9&size=600x400&markers=icon:large-blue-cutout%7C${latitude},${longitude}`
+  };
 
-  }
-}
+  // Function to handle errors
+  const handleLocationError = (error) => {
+    console.error('Error fetching location info:', error);
+    setErrorResponse(error.message);
+    setErrorResponseBody(error.response?.data?.error || 'An unknown error occurred');
+    toggleModal();
+  };
+
+  const toggleModal = () => {
+    setShow((prevShow) => !prevShow);
+  };
+
+  const hasValidCityData = () => {
+    return cityName && latitude && longitude && mapImageUrl; // return true only if all values are truthy
+  };
+
+  return (
+    <>
+      <div className='form-container'>
+        <Form onSubmit={handleFormSubmit}>
+          <Form.Group className="mb-3" controlId="formBasicEmail">
+            <Form.Label>Search</Form.Label>
+            <Form.Control 
+              type="text" 
+              placeholder="Enter name of city here"
+              value={userInput}
+              onChange={handleUserInput} 
+            />
+          </Form.Group>
+
+          <Button variant="primary" type="submit" className='mb-3'>
+            Explore!
+          </Button>
+        </Form>
+      </div>
+
+      <div className='card-container'>
+        {hasValidCityData() && ( // render city info only if valid city data 
+          <CityInfo 
+            cityName={cityName} 
+            latitude={latitude} 
+            longitude={longitude} 
+            mapImageUrl={mapImageUrl}
+          />
+        )}
+      </div>
+
+      <ErrorModal 
+        errorTitle={errorResponse} 
+        errorBody={errorResponseBody} 
+        show={show}
+        toggleModal={toggleModal}
+      />
+    </>
+  );
+};
 
 export default Explorer;
+
+// class UserForm extends React.Component {
+//   constructor(props) {
+//     super(props);
+
+//     // Initialize states
+//     this.state = {
+//       searchQuery: '', // User input
+//       location: null, // API location data based on query
+//       staticMapURL: null, // Static map url based on location
+//       showErrorModal: false, // Show error modal when true
+//       errorMessage: '', // Show 4xx error message in modal title
+//       errorMessageBody: '', // Message to show in body of modal
+//       serverResponseData: [], // Holds forecast data from server
+//       getForecastDataError: false, // Track errors in getForecastData()
+//       serverMovieResponseData: [],
+//     };
+//   }
+
+
+//   // Function to handle form submission
+//   handleForm = (event) => {
+    
+//     event.preventDefault();
+
+//     const { searchQuery } = this.state;
+
+//     // Make API requests
+//     const us1Url = `https://us1.locationiq.com/v1/search?key=${API_KEY}&q=${searchQuery}&format=json`;
+//     const eu1Url = `https://eu1.locationiq.com/v1/search?key=${API_KEY}&q=${searchQuery}&format=json`;
+
+//     // Promise.all() - allows for multiple promises to complete and then proceed with 'then' processing once all promises resolved
+//     Promise.all([
+//       this.makeApiRequest(us1Url), 
+//       this.makeApiRequest(eu1Url)
+//     ])
+//       .then( () => {
+        
+//         if (this.state.showErrorModal) { // If showErrorModal true, breakout of function
+//           return;
+//         } else { // Else get forecast data
+//           this.getForecastData() // Make call to server using query data
+//             .then(() => {
+              
+//               if (this.state.showErrorModal) {
+//                 return;
+//               } else {
+//                 this.getMovieData(); // Make call to server to get movie data
+//               }
+//             });
+//         }
+//       });
+//   };
+
+//   // Function to make API request and handle response/error
+//     makeApiRequest = (url) => {
+//       return axios.get(url)
+//         .then((response) => {
+//           this.setState({ location: response.data[0] }); // Set state of location to first element in response.data array
+//           this.updateStaticMapURL();
+
+//         })
+//         // Catch error and toggle errorModal
+//         .catch((error) => {
+//           this.setState({ 
+//             errorMessage: error.message, 
+//             errorMessageBody: error.response.data.error, 
+//           }); // Capture error message from console log and set to errorMessage variable
+
+//           // Check if showErrorModal is already true before toggling it (since passing in two urls)
+//           if (!this.state.showErrorModal) {
+//             this.toggleErrorModal();
+//           }
+
+//         });
+//     };
+
+
+//   // Function - retrieve forecast data from server
+//   getForecastData = async () => {
+
+//     const { location } = this.state; // Access location property/data of this.state.location
+
+//     try {
+//       const response = await axios.get(`${SERVER}/weather?lat=${location.lat}&lon=${location.lon}`);
+
+//       const weather = response.data.data;
+//       const weatherTimeStamp = response.data.timestamp;
+      
+//       // If the request is successful, update the state with the response data
+//       this.setState({ 
+//         serverResponseData: weather,
+//         serverWeatherTimeStamp: weatherTimeStamp, 
+//         errorMessage: '' });
+
+//     } catch (error) {
+//       // If there's an error, catch it and set errorMessage state
+//       console.error('Server Error:', error);
+
+//       this.setState({ 
+//         errorMessage: error.message, 
+//         serverResponseData: [],
+//         errorMessageBody: error.response.data.error, 
+//         getForecastDataError: true, // Set getForecastDataError to true
+//        });
+
+//        this.toggleErrorModal();
+
+//     }
+
+//   };
+
+
+//   // Function - retrieve movie data from server
+//   getMovieData = async () => {
+
+//     const { location } = this.state; // Access location property/data of this.state.location
+//     let displayName = location.display_name; // Retrieve display name of city (City,County, State)
+//     let cityNameOnly = displayName.split(',')[0].trim(); // Isolate name of city only, (first word before a comma)
+
+
+//     try {
+//       const response = await axios.get(`${SERVER}/movies?searchQuery=${cityNameOnly}`);
+
+//       const movies = response.data.data;
+//       const movieTimeStamp = response.data.timestamp;
+      
+//       // If the request is successful, update the state with the response data
+//       this.setState({ 
+//         serverMovieResponseData: movies, 
+//         serverMovieTimestamp: movieTimeStamp,
+//         errorMessage: '' });
+
+//     } catch (error) {
+//       // If there's an error, catch it and set errorMessage state
+//       console.error('Server Error:', error);
+
+//       this.setState({ 
+//         errorMessage: error.message, 
+//         serverResponseData: [],
+//         errorMessageBody: error.response.data.error, 
+//         getForecastDataError: true, // Set getForecastDataError to true
+//        });
+
+//        this.toggleErrorModal();
+
+//     }
+
+//   };
+  
+
+//   // Function to update staticMapURL based on location
+//   updateStaticMapURL() {
+//     const { location } = this.state; // Access location property/data of this.state.location
+//     if (location) { // If location is truthy, assign latitude and longitude of location to staticMapURL
+//       this.setState({
+//         staticMapURL: `https://maps.locationiq.com/v3/staticmap?key=${API_KEY}&center=${location.lat},${location.lon}&zoom=9&size=600x600&markers=icon:large-blue-cutout%7C${location.lat},${location.lon}`,
+//       });
+//     }
+//   }
+
+
+//   // Update state of query to value of input field (input by user)
+//   handleChange = (event) => {
+//     this.setState({ searchQuery: event.target.value });
+//   };
+
+
+//   // Function to toggle value of showErrorModal (true and false)
+//   toggleErrorModal = () => {
+//     this.setState((prevState) => ({
+//       showErrorModal: !prevState.showErrorModal,
+//     }));
+//   };
+
+
+//   // Function to close modal
+//   closeModal = () => {
+//     this.setState({
+//       showErrorModal: false,
+//       getForecastDataError: false,
+//     });
+//   };
+  
+
+//   render() {
+//     return (
+//       <>
+//         {/* Run handleForm when user clicks Explore! button */}
+//         <Form onSubmit={this.handleForm}>
+
+//           <Form.Group className="mb-3">
+//             <Form.Label>
+//               Search 
+//             </Form.Label>
+
+//             <Form.Control
+//               onChange={this.handleChange}
+//               name="city"
+//               type="text"
+//               placeholder="Enter name of city here"
+//             />
+//           </Form.Group>
+
+//           <Button 
+//             variant="primary" 
+//             type="submit">
+//               Explore!
+//           </Button>
+
+//         </Form>
+
+//         {/* Show modal when error caught, toggle error modal to false when modal closes */}
+//         <Modal 
+//           show={this.state.showErrorModal} 
+//           onHide={this.toggleErrorModal} 
+//           className="modal">
+
+//           <Modal.Header closeButton>
+//             <Modal.Title>
+//               {this.state.errorMessage}
+//             </Modal.Title>
+//           </Modal.Header>
+
+//           <Modal.Body>
+//               {this.state.errorMessageBody}
+//           </Modal.Body>
+
+//           <Modal.Footer>
+//             <Button 
+//               variant="secondary" 
+//               onClick={this.toggleErrorModal}>
+//               Close
+//             </Button>
+//           </Modal.Footer>
+
+//         </Modal>
+
+//         <Explorer
+//           displayName={this.state.location && this.state.location.display_name}
+//           latitude={this.state.location && this.state.location.lat}
+//           longitude={this.state.location && this.state.location.lon}
+//           staticMapURL={this.state.location && this.state.staticMapURL}
+//         />
+
+//         <Weather 
+        
+//           serverResponseData={this.state.serverResponseData && this.state.serverResponseData}
+//           serverWeatherTimeStamp = {this.state.serverResponseData && this.state.serverWeatherTimeStamp}
+        
+//         />
+
+//         <Movies 
+        
+//           serverMovieResponseData = {this.state.serverMovieResponseData && this.state.serverMovieResponseData}
+//           serverMovieTimestamp = {this.state.serverMovieResponseData && this.state.serverMovieTimestamp}
+//           cityName = {this.state.location && this.state.location.display_name}
+        
+//         />
+
+//       </>
+//     );
+//   }
+// }
